authentication and authorization is complete in journal app

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -1,7 +1,10 @@
 package net.engineeringdigest.journalApp.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
+
+import javax.management.RuntimeErrorException;
 
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,21 +28,16 @@ public class JournalEntryService {
 	
 	@Transactional
 	public void saveEntry(JournalEntry journalEntry, String userName) {
-
+		try {
 	    User user = userService.findByUserName(userName);
-
-	    if (user == null) {
-	        throw new RuntimeException("User not found");
-	    }
-
-	    // save journal entry
+	    journalEntry.setDate(LocalDateTime.now());
 	    JournalEntry saved = journalEntryRepository.save(journalEntry);
-
-	    // add entry to user
 	    user.getJournalEntries().add(saved);
-
-	    // save updated user
 	    userService.saveUser(user);
+		}catch (Exception e) {
+			System.out.println(e);
+			throw new RuntimeException("An Error Ocuured while saving the entry",e);
+		}
 	}
 	
 	public void saveEntry(JournalEntry journalEntry) {
@@ -54,12 +52,25 @@ public class JournalEntryService {
 		return journalEntryRepository.findById(id);
 	}
 	
-	public void deleteById(ObjectId id ,String userName) {
-		User user = userService.findByUserName(userName);
-		user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-		userService.saveUser(user);
-		journalEntryRepository.deleteById(id);
+	@Transactional
+	public boolean deleteById(ObjectId id ,String userName) {
+		boolean removed = false;
+		try {
+			User user = userService.findByUserName(userName);
+			removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+			if(removed) {
+				userService.saveUser(user);
+				journalEntryRepository.deleteById(id);
+			}
+		}catch (Exception e) {
+			System.out.println(e);
+			throw new RuntimeException("An error ocuured while deleting the entry");
+		}
+		return removed;
+		
 	}
+	
+	
 }
 
 // Controller ---> service ----> repository
